Creo documentación en readme e implemento carpeta img con archivos de diagramas

+++ src/main/java/com/entidades/Atraccion.java
@@ -20,7 +20,7 @@ public class Atraccion {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false)
     private String nombre;
@@ -56,13 +56,16 @@ public class Atraccion {
 
 
 
-	public Integer getId() {
+
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+
+	public void setId(Long id) {
 		this.id = id;
 	}
+
 
 	public String getNombre() {
 		return nombre;
